+++ lib/bloc_effects.dart
@@ -4,3 +4,4 @@ export '/src/bloc_effect_listener.dart';
 export '/src/bloc_with_effects.dart';
 export '/src/bloc_with_effects_observer.dart';
 export '/src/cubit_with_effects.dart';
+export '/src/effects_mixin.dart';