+++ lib/src/bloc_with_effects.dart
@@ -1,4 +1,4 @@
-import 'package:bloc_effects/src/base_effector.dart';
+import 'package:bloc_effects/src/effects_mixin.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 /// {@template bloc}
@@ -6,8 +6,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 /// and add the [emitEffect] method to [Bloc] features.
 /// {@endtemplate}
 abstract class BlocWithEffects<Event, State, Effect> extends Bloc<Event, State>
-    with Effector<State, Effect>
-    implements BaseEffector<Effect> {
+    with Effects<Effect> {
   /// {@macro bloc}
   BlocWithEffects(super.initialState);
 }