+++ test/bloc_with_effects_observer_test.dart
@@ -6,13 +6,11 @@ import 'bloc_effect_listener_test.dart';
 class TestBlocObserver extends BlocWithEffectsObserver {}
 
 void main() {
-  final cubit = TestCubit();
-
   group('BlocWithEffectsObserver', () {
     group('onEffect', () {
       test('does nothing by default', () {
         // ignore: invalid_use_of_protected_member
-        TestBlocObserver().onEffect(cubit, const ShowSnackBar());
+        TestBlocObserver().onEffect(const ShowSnackBar(1));
       });
     });
   });